Adding new type of output - Javascript Object
I love the fact that the MySQL output can be pasted into SQL editor. I needed the same thing for pasting into Javascript editor when dealing with really large spreadsheets.
The code changes in this commit are simply the addition of native instantiation of a Javascript object named mrDataConverterObj, the removal of the newline character (after each record) and enclosing the entire output with single quotes (as you would do if you wanted to parse the output as a JSON object in Javascript).